mcp-client-python/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.5",
    "mcp>=1.1.1",
    "python-dotenv>=1.0.1",
]
